==> 2_efficient_data_format_preprocess.ipynb <==
### 將從 kaggle 下載下來有 label csv 檔案進行前處理，以符合 EfficientNet 訓練的格式

import os
import glob
import numpy as np
import pandas as pd
from config import Args

train = pd.read_csv(os.path.join(Args.RSNA_ROOT_DIR, 'stage_2_train.csv'))
# test = pd.read_csv(os.path.join(Args.RSNA_ROOT_DIR, 'stage_2_sample_submission.csv'))

train[['ID', 'Image', 'Diagnosis']] = train['ID'].str.split('_', expand=True)
train = train[['Image', 'Diagnosis', 'Label']]
train.drop_duplicates(inplace=True)
train = train.pivot(index='Image', columns='Diagnosis', values='Label').reset_index()
train['Image'] = 'ID_' + train['Image']
train.head()


# Some files didn't contain legitimate images, so we need to remove them

png = glob.glob(os.path.join(Args.ALL_PNG_IMAGE, '*.png'))
png = [os.path.basename(png)[:-4] for png in png]
png = np.array(png)

train = train[train['Image'].isin(png)]

train.to_csv(os.path.join(Args.TRAIN_CSV_PATH, 'train.csv'), index=False)


# Also prepare the test data

# test[['ID','Image','Diagnosis']] = test['ID'].str.split('_', expand=True)
# test['Image'] = 'ID_' + test['Image']
# test = test[['Image', 'Label']]
# test.drop_duplicates(inplace=True)

# test.to_csv('test.csv', index=False)
