==> 4_cp_image.ipynb <==
## 將 site 的圖片資料移出

import pandas as pd
import shutil
import os
from config import Args

dirs = os.listdir(Args.FL_SITE_CSV_PATH)
count = 0
for i, file in enumerate(dirs):
    file_name = file.split(".")
    csv_file = os.path.join(Args.FL_SITE_CSV_PATH, file)
    output_folder = os.path.join(Args.FL_SITE_IMAGE_PATH, f"{file_name[0]}-images")  # 指定的資料夾

    # 確保輸出資料夾存在，如果不存在就創建
    os.makedirs(output_folder, exist_ok=True)

    # 讀取CSV檔案
    df = pd.read_csv(csv_file)

    # 遍歷CSV中的每一行，並處理Image欄位
    for index, row in df.iterrows():
        image_id = row["Image"]
        image_path = os.path.join(Args.ALL_PNG_IMAGE, f"{image_id}.png")

        # 如果在train_png資料夾中找到相應的PNG檔案，則複製到輸出資料夾
        if os.path.exists(image_path):
            output_path = os.path.join(output_folder, f"{image_id}.png")
            shutil.copy(image_path, output_path)
            print(f"複製檔案 {image_id}.png 到 {output_folder}")
        else:
            print(f"找不到檔案 {image_id}.png")

    print("完成複製檔案。")

### 將 test 的圖片資料移出

csv_file = os.path.join(Args.FL_TEST_CSV_PATH, "test_data.csv")
output_folder = os.path.join(Args.FL_SITE_IMAGE_PATH, "test-images")  # 指定的資料夾

# 確保輸出資料夾存在，如果不存在就創建
os.makedirs(output_folder, exist_ok=True)

# 讀取CSV檔案
df = pd.read_csv(csv_file)

# 遍歷CSV中的每一行，並處理Image欄位
for index, row in df.iterrows():
    image_id = row["Image"]
    image_path = os.path.join(Args.ALL_PNG_IMAGE, f"{image_id}.png")

    # 如果在train_png資料夾中找到相應的PNG檔案，則複製到輸出資料夾
    if os.path.exists(image_path):
        output_path = os.path.join(output_folder, f"{image_id}.png")
        shutil.copy(image_path, output_path)
        print(f"複製檔案 {image_id}.png 到 {output_folder}")
    else:
        print(f"找不到檔案 {image_id}.png")

print("完成複製檔案。")
